# backend/internal/gen/http/openapi.yaml
swagger: "2.0"
info:
  title: Wargame Service
  description: HTTP service for the SSM Wargame
  version: ""
host: localhost:80
consumes:
- application/json
- application/xml
- application/gob
produces:
- application/json
- application/xml
- application/gob
paths:
...
  /admin/challenges:
    get:
      tags:
      - admin
      summary: ListChallenges admin
      operationId: admin#ListChallenges
      parameters:
      - name: Authorization
        in: header
        required: true
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/AdminSsmAdminChallengeResponseCollection'
        "400":
          description: Resource not found
          schema:
            $ref: '#/definitions/AdminListChallengesBadRequestResponseBody'
        "403":
          description: When the user does not have the required role
          schema:
            $ref: '#/definitions/AdminListChallengesUnauthorizedResponseBody'
        "404":
          description: Resource not found
          schema:
            $ref: '#/definitions/AdminListChallengesNotFoundResponseBody'
      schemes:
      - http
      security:
      - jwt_header_Authorization: []
    post:
      tags:
      - admin
      summary: CreateChallenge admin
      operationId: admin#CreateChallenge
      parameters:
      - name: Authorization
        in: header
        required: true
        type: string
      - name: CreateChallengeRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/AdminCreateChallengeRequestBody'
          required:
          - category_id
          - id
          - title
          - description
          - score
          - slug
          - solves
          - category
      responses:
        "201":
          description: Created response.
        "400":
          description: Resource not found
          schema:
            $ref: '#/definitions/AdminCreateChallengeBadRequestResponseBody'
        "403":
          description: When the user does not have the required role
          schema:
            $ref: '#/definitions/AdminCreateChallengeUnauthorizedResponseBody'
        "404":
          description: Resource not found
          schema:
            $ref: '#/definitions/AdminCreateChallengeNotFoundResponseBody'
      schemes:
      - http
      security:
      - jwt_header_Authorization: []
...
  /admin/monthly_challenges:
    get:
      tags:
      - admin
      summary: ListMonthlyChallenges admin
      operationId: admin#ListMonthlyChallenges
      parameters:
      - name: Authorization
        in: header
        required: true
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            type: array
            items:
              $ref: '#/definitions/MonthlyChallengeResponse'
        "400":
          description: Resource not found
          schema:
            $ref: '#/definitions/AdminListMonthlyChallengesBadRequestResponseBody'
        "403":
          description: When the user does not have the required role
          schema:
            $ref: '#/definitions/AdminListMonthlyChallengesUnauthorizedResponseBody'
        "404":
          description: Resource not found
          schema:
            $ref: '#/definitions/AdminListMonthlyChallengesNotFoundResponseBody'
      schemes:
      - http
      security:
      - jwt_header_Authorization: []
    post:
      tags:
      - admin
      summary: CreateMonthlyChallenge admin
      operationId: admin#CreateMonthlyChallenge
      parameters:
      - name: Authorization
        in: header
        required: true
        type: string
      - name: CreateMonthlyChallengeRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/AdminCreateMonthlyChallengeRequestBody'
          required:
          - start_date
          - end_date
          - display_month
          - challenge_id
          - id
          - title
          - description
          - score
          - slug
          - solves
          - category
      responses:
        "204":
          description: No Content response.
        "400":
          description: Resource not found
          schema:
            $ref: '#/definitions/AdminCreateMonthlyChallengeBadRequestResponseBody'
        "403":
          description: When the user does not have the required role
          schema:
            $ref: '#/definitions/AdminCreateMonthlyChallengeUnauthorizedResponseBody'
        "404":
          description: Resource not found
          schema:
            $ref: '#/definitions/AdminCreateMonthlyChallengeNotFoundResponseBody'
      schemes:
      - http
      security:
      - jwt_header_Authorization: []
...
  /challenges:
    get:
      tags:
      - challenge
      summary: ListChallenges challenge
      operationId: challenge#ListChallenges
      parameters:
      - name: Authorization
        in: header
        required: false
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/ChallengeSsmChallengeResponseCollection'
      schemes:
      - http
      security:
      - jwt_header_Authorization: []
  /challenges/{challengeID}/attempt:
    post:
      tags:
      - challenge
      summary: SubmitFlag challenge
      operationId: challenge#SubmitFlag
      parameters:
      - name: challengeID
        in: path
        description: ID of a challenge
        required: true
        type: string
        format: uuid
      - name: Authorization
        in: header
        required: true
        type: string
      - name: SubmitFlagRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/ChallengeSubmitFlagRequestBody'
          required:
          - flag
      responses:
        "200":
          description: OK response.
        "400":
          description: If the challenge is already solved
          schema:
            $ref: '#/definitions/ChallengeSubmitFlagIncorrectFlagResponseBody'
        "409":
          description: If the challenge is already solved
          schema:
            $ref: '#/definitions/ChallengeSubmitFlagAlreadySolvedResponseBody'
      schemes:
      - http
      security:
      - jwt_header_Authorization: []
  /monthly_challenges:
    get:
      tags:
      - challenge
      summary: ListMonthlyChallenges challenge
      operationId: challenge#ListMonthlyChallenges
      parameters:
      - name: Authorization
        in: header
        required: false
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            type: array
            items:
              $ref: '#/definitions/MonthlyChallengeResponse'
      schemes:
      - http
      security:
      - jwt_header_Authorization: []
...
definitions:
...
  AdminCreateChallengeBadRequestResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: false
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: true
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: false
    description: CreateChallenge_bad_request_Response_Body result type (default view)
    example:
      fault: false
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: false
      timeout: false
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  AdminCreateChallengeNotFoundResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: true
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: true
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: false
    description: CreateChallenge_not_found_Response_Body result type (default view)
    example:
      fault: true
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: true
      timeout: true
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  AdminCreateChallengeRequestBody:
    title: AdminCreateChallengeRequestBody
    type: object
    properties:
      category_id:
        type: string
        example: 1b678292-6737-4cc7-8eae-aa821100293f
        format: uuid
      ctf_event_id:
        type: string
        description: The ID of the CTF the challenge was taken from
        example: c397efb2-b171-4d77-9166-d105cf4f521a
        format: uuid
      description:
        type: string
        description: A short text describing the challenge
        example: A heap overflow challenge
      publish_at:
        type: integer
        description: unix timestamp
        example: 1638384718
        format: int64
      score:
        type: integer
        description: The number of points given to the solver
        example: 50
        format: int64
      slug:
        type: string
        description: A unique string that can be used in URLs
        example: pwnme
      title:
        type: string
        description: Title displayed to user
        example: pwnme
    example:
      category_id: 1b678292-6737-4cc7-8eae-aa821100293f
      ctf_event_id: c397efb2-b171-4d77-9166-d105cf4f521a
      description: A heap overflow challenge
      publish_at: 1638384718
      score: 50
      slug: pwnme
      title: pwnme
    required:
    - category_id
    - id
    - title
    - description
    - score
    - slug
    - solves
    - category
  AdminCreateChallengeUnauthorizedResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: true
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: false
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: false
    description: CreateChallenge_unauthorized_Response_Body result type (default view)
    example:
      fault: false
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: false
      timeout: false
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
...
  AdminListChallengesBadRequestResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: false
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: true
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: true
    description: ListChallenges_bad_request_Response_Body result type (default view)
    example:
      fault: false
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: false
      timeout: true
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  AdminListChallengesNotFoundResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: false
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: false
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: true
    description: ListChallenges_not_found_Response_Body result type (default view)
    example:
      fault: true
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: true
      timeout: true
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  AdminListChallengesUnauthorizedResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: true
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: true
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: false
    description: ListChallenges_unauthorized_Response_Body result type (default view)
    example:
      fault: true
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: false
      timeout: false
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  AdminListMonthlyChallengesBadRequestResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: false
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: false
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: false
    description: ListMonthlyChallenges_bad_request_Response_Body result type (default
      view)
    example:
      fault: true
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: false
      timeout: false
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
...
  AdminListMonthlyChallengesUnauthorizedResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: false
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: false
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: true
    description: ListMonthlyChallenges_unauthorized_Response_Body result type (default
      view)
    example:
      fault: true
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: true
      timeout: false
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
...
  AdminSsmAdminChallengeResponseCollection:
    title: 'Mediatype identifier: application/vnd.ssm.admin.challenge; type=collection;
      view=default'
    type: array
    items:
      $ref: '#/definitions/SsmAdminChallengeResponse'
    description: ListChallengesResponseBody is the result type for an array of SsmAdminChallengeResponse
      (default view)
    example:
    - category_id: 12b8dc3a-10ae-49ed-9d69-5208ccd92ed1
      description: A heap overflow challenge
      files:
      - bucket: ssm-prod-bucket
        filename: decryptor.exe
        id: 71333e34-4c6b-483e-b3c7-c77d73008cca
        key: file.exe
        md5: cq02dBbcuugBHM1oKyvMlQ==
        size: 87123
        url: https://s3/bucket/key?signature
      - bucket: ssm-prod-bucket
        filename: decryptor.exe
        id: 71333e34-4c6b-483e-b3c7-c77d73008cca
        key: file.exe
        md5: cq02dBbcuugBHM1oKyvMlQ==
        size: 87123
        url: https://s3/bucket/key?signature
      - bucket: ssm-prod-bucket
        filename: decryptor.exe
        id: 71333e34-4c6b-483e-b3c7-c77d73008cca
        key: file.exe
        md5: cq02dBbcuugBHM1oKyvMlQ==
        size: 87123
        url: https://s3/bucket/key?signature
      - bucket: ssm-prod-bucket
        filename: decryptor.exe
        id: 71333e34-4c6b-483e-b3c7-c77d73008cca
        key: file.exe
        md5: cq02dBbcuugBHM1oKyvMlQ==
        size: 87123
        url: https://s3/bucket/key?signature
      flags:
      - flag: SSM{yo}
        id: 71333e34-4c6b-483e-b3c7-c77d73008cca
      - flag: SSM{yo}
        id: 71333e34-4c6b-483e-b3c7-c77d73008cca
      id: e721a338-44de-4de8-a562-43d2db5f4115
      publish_at: 1638384718
      score: 50
      services:
      - {}
      - {}
      - {}
      slug: pwnme
      solves: 3
      title: pwnme
    - category_id: 12b8dc3a-10ae-49ed-9d69-5208ccd92ed1
      description: A heap overflow challenge
      files:
      - bucket: ssm-prod-bucket
        filename: decryptor.exe
        id: 71333e34-4c6b-483e-b3c7-c77d73008cca
        key: file.exe
        md5: cq02dBbcuugBHM1oKyvMlQ==
        size: 87123
        url: https://s3/bucket/key?signature
      - bucket: ssm-prod-bucket
        filename: decryptor.exe
        id: 71333e34-4c6b-483e-b3c7-c77d73008cca
        key: file.exe
        md5: cq02dBbcuugBHM1oKyvMlQ==
        size: 87123
        url: https://s3/bucket/key?signature
      - bucket: ssm-prod-bucket
        filename: decryptor.exe
        id: 71333e34-4c6b-483e-b3c7-c77d73008cca
        key: file.exe
        md5: cq02dBbcuugBHM1oKyvMlQ==
        size: 87123
        url: https://s3/bucket/key?signature
      - bucket: ssm-prod-bucket
        filename: decryptor.exe
        id: 71333e34-4c6b-483e-b3c7-c77d73008cca
        key: file.exe
        md5: cq02dBbcuugBHM1oKyvMlQ==
        size: 87123
        url: https://s3/bucket/key?signature
      flags:
      - flag: SSM{yo}
        id: 71333e34-4c6b-483e-b3c7-c77d73008cca
      - flag: SSM{yo}
        id: 71333e34-4c6b-483e-b3c7-c77d73008cca
      id: e721a338-44de-4de8-a562-43d2db5f4115
      publish_at: 1638384718
      score: 50
      services:
      - {}
      - {}
      - {}
      slug: pwnme
      solves: 3
      title: pwnme
    - category_id: 12b8dc3a-10ae-49ed-9d69-5208ccd92ed1
      description: A heap overflow challenge
      files:
      - bucket: ssm-prod-bucket
        filename: decryptor.exe
        id: 71333e34-4c6b-483e-b3c7-c77d73008cca
        key: file.exe
        md5: cq02dBbcuugBHM1oKyvMlQ==
        size: 87123
        url: https://s3/bucket/key?signature
      - bucket: ssm-prod-bucket
        filename: decryptor.exe
        id: 71333e34-4c6b-483e-b3c7-c77d73008cca
        key: file.exe
        md5: cq02dBbcuugBHM1oKyvMlQ==
        size: 87123
        url: https://s3/bucket/key?signature
      - bucket: ssm-prod-bucket
        filename: decryptor.exe
        id: 71333e34-4c6b-483e-b3c7-c77d73008cca
        key: file.exe
        md5: cq02dBbcuugBHM1oKyvMlQ==
        size: 87123
        url: https://s3/bucket/key?signature
      - bucket: ssm-prod-bucket
        filename: decryptor.exe
        id: 71333e34-4c6b-483e-b3c7-c77d73008cca
        key: file.exe
        md5: cq02dBbcuugBHM1oKyvMlQ==
        size: 87123
        url: https://s3/bucket/key?signature
      flags:
      - flag: SSM{yo}
        id: 71333e34-4c6b-483e-b3c7-c77d73008cca
      - flag: SSM{yo}
        id: 71333e34-4c6b-483e-b3c7-c77d73008cca
      id: e721a338-44de-4de8-a562-43d2db5f4115
      publish_at: 1638384718
      score: 50
      services:
      - {}
      - {}
      - {}
      slug: pwnme
      solves: 3
      title: pwnme
    - category_id: 12b8dc3a-10ae-49ed-9d69-5208ccd92ed1
      description: A heap overflow challenge
      files:
      - bucket: ssm-prod-bucket
        filename: decryptor.exe
        id: 71333e34-4c6b-483e-b3c7-c77d73008cca
        key: file.exe
        md5: cq02dBbcuugBHM1oKyvMlQ==
        size: 87123
        url: https://s3/bucket/key?signature
      - bucket: ssm-prod-bucket
        filename: decryptor.exe
        id: 71333e34-4c6b-483e-b3c7-c77d73008cca
        key: file.exe
        md5: cq02dBbcuugBHM1oKyvMlQ==
        size: 87123
        url: https://s3/bucket/key?signature
      - bucket: ssm-prod-bucket
        filename: decryptor.exe
        id: 71333e34-4c6b-483e-b3c7-c77d73008cca
        key: file.exe
        md5: cq02dBbcuugBHM1oKyvMlQ==
        size: 87123
        url: https://s3/bucket/key?signature
      - bucket: ssm-prod-bucket
        filename: decryptor.exe
        id: 71333e34-4c6b-483e-b3c7-c77d73008cca
        key: file.exe
        md5: cq02dBbcuugBHM1oKyvMlQ==
        size: 87123
        url: https://s3/bucket/key?signature
      flags:
      - flag: SSM{yo}
        id: 71333e34-4c6b-483e-b3c7-c77d73008cca
      - flag: SSM{yo}
        id: 71333e34-4c6b-483e-b3c7-c77d73008cca
      id: e721a338-44de-4de8-a562-43d2db5f4115
      publish_at: 1638384718
      score: 50
      services:
      - {}
      - {}
      - {}
      slug: pwnme
      solves: 3
      title: pwnme
...
  ChallengeSubmitFlagRequestBody:
    title: ChallengeSubmitFlagRequestBody
    type: object
    properties:
      flag:
        type: string
        example: SSM{flag}
        maxLength: 200
    example:
      flag: SSM{flag}
    required:
    - flag
  MonthlyChallengeResponse:
    title: MonthlyChallengeResponse
    type: object
    properties:
      challenge_id:
        type: string
        example: 85163218-8735-42ed-a7a6-42a9de2294df
        format: uuid
      description:
        type: string
        description: A short text describing the challenge
        example: A heap overflow challenge
      display_month:
        type: string
        description: The month(s) that the challenge is assigned for
        example: Januari/Februari
      end_date:
        type: string
        description: Ending date of the monthly challenge
        example: "2006-02-01"
      slug:
        type: string
        description: A unique string that can be used in URLs
        example: pwnme
      start_date:
        type: string
        description: Starting date of the monthly challenge
        example: "2006-02-01"
      title:
        type: string
        description: Title displayed to user
        example: pwnme
    example:
      challenge_id: 85163218-8735-42ed-a7a6-42a9de2294df
      description: A heap overflow challenge
      display_month: Januari/Februari
      end_date: "2006-02-01"
      slug: pwnme
      start_date: "2006-02-01"
      title: pwnme
    required:
    - start_date
    - end_date
    - display_month
    - challenge_id
    - id
    - title
    - description
    - score
    - slug
    - solves
    - category
...
securityDefinitions:
  jwt_header_Authorization:
    type: apiKey
    name: Authorization
    in: header